Improve logging, Disable broken Winston log rotation
Winston creates log files with a number suffix,
but the most recent log file can have any number.
This makes it hard to find the actual log file.

Use the system's logrotate feature instead.

Load configuration as a JSON file, not as a module.

Protect logger against missing configuration.

Rename log file from debug.js to server.js as it's
not related to the log level.