Introduce method to write album index file
When overwriting an index file, ensure that any additional keys in the
old index file are also included in the new file.
Use `json-stable-stringify` module to ensure index members are written
in a stable, predefined order. `JSON.stringify` accepts an array of
keys to define the order, but this would skip all unknown keys.