Fix issues with overwritten paths and tags in index
Ensure that track paths are stored relative to the album path in
index.json. Path in the model are relative to the music dir.

Also ensure that tags are removed if there are tags in the old
index but not in the new one.

Ensure that only albums with a valid path are added.
Same for cover images.