Revise logging, replace winston with bunyan
Replace `winston` with `bunyan`, which seems simpler and creates
machine-readable (JSON) logs. Wrap the bunyan API with a simplified
logger API to reduce coupling.

Rewrite logging statements to use template strings for messages and
to pass errors in an `err` field.

Write HTTP access log to a separate file, `access.log`. It shouldn't
be mixed with server logs.

Enable log rotation using `rotating-file-stream`.

Add a config property `logToConsole` and set it in the `config.json`
that is used only for development.