Remove frontend, move REST API to /api
Entirely remove the current frontend to make way for a new one that
is to be developed separately. Expect the frontend's static files to
be copied into a `/static/` folder. Deliver any static files that are
present in this folder but do not provide any default static content.

Remove the `/views` and `/statics` folders.

Move the REST API to `/api/`. Remove all code that renders HTML.